auth middleware

+++ lib/actions/user.action.ts
@@ -5,82 +5,86 @@ import { connectToDatabase } from "../mongoose";
 import {
   CreateUserParams,
   DeleteUserParams,
-  GetUserByIdParams,
   UpdateUserParams,
 } from "./shared.types";
 import { revalidatePath } from "next/cache";
 import Question from "@/database/question.modal";
 
-export async function getUserById(params: GetUserByIdParams) {
+export const getUserById = async (params: any) => {
   try {
     connectToDatabase();
     const { userId } = params;
-    const user = await User.findOne({
-      clearId: userId,
-    });
+
+    // get user by id
+    const user = await User.findOne({ clerkId: userId });
+
     return user;
   } catch (error) {
     console.log(error);
     throw error;
   }
-}
+};
 
-export async function createUser(userData: CreateUserParams) {
+// create a user in the database
+export const createUser = async (userData: CreateUserParams) => {
   try {
     connectToDatabase();
     const newUser = await User.create(userData);
+
     return newUser;
   } catch (error) {
     console.log(error);
     throw error;
   }
-}
+};
 
-export async function updateUser(params: UpdateUserParams) {
+export const updateUser = async (params: UpdateUserParams) => {
   try {
     connectToDatabase();
+
     const { clerkId, updateData, path } = params;
-    await User.findOneAndUpdate(
-      {
-        clerkId,
-      },
-      updateData,
-      {
-        new: true,
-      }
-    );
+
+    await User.findOneAndUpdate({ clerkId }, updateData, {
+      new: true,
+    });
+
     revalidatePath(path);
   } catch (error) {
     console.log(error);
     throw error;
   }
-}
+};
 
-export async function deleteUser(params: DeleteUserParams) {
+export const deleteUser = async (params: DeleteUserParams) => {
   try {
     connectToDatabase();
-    const { clerkId } = params;
 
+    const { clerkId } = params;
     const user = await User.findOneAndDelete({ clerkId });
     if (!user) {
       throw new Error("User not found");
     }
-    // Delete user from database
-    // and questions, answers, comments, etc.
+
+    // if user exist delete user data from database
+    // quesions,answers,comments etc
+
+    // get user question ids
+
     // const userQuestionIds = await Question.find({
-    //   author: user._id,
-    // }).distinct("_id");
+    //   author: user._id
+    // }).distinct('_id');
 
-    // console.log(userQuestionIds);
+    // delete user Questions
+    await Question.deleteMany({ author: user._id });
 
-    await Question.deleteMany({
-      author: user._id,
-    });
+    // todo: delete user answers, comments etc
 
-    const deleteUser = await User.findByIdAndDelete(user._id);
-    return deleteUser;
+    // delete user
+    const deletedUser = await User.findByIdAndDelete(user._id);
+
+    return deletedUser;
   } catch (error) {
     console.log(error);
     throw error;
   }
-}
+};
